requirements.txt
# requirements.txt
tensorflow          
scikit-learn
pandas
numpy
matplotlib
seaborn
streamlit
streamlit-webrtc
opencv-python      
requests
tqdm
Pillowgit   